Changed var to env variable

+++ LintConfig.sh
@@ -1,6 +1,14 @@
 workingDir=$(pwd)
-projectDir="/Users/nibe/git/wireframe_linter"
+projectDir=$LINTER_DIR
 
-mkdir $projectDir/temp_files
-node $projectDir/index.js $workingDir
-rm -rf $projectDir/temp_files
+
+if [ -z "$LINTER_DIR" ]; then
+	echo '---------------------------------------------------'
+	echo 'Environment variable LINTER_DIR not provided'
+	echo 'Please export LINTER_DIR=<path_of_wireframe_linter>'
+	echo '---------------------------------------------------'
+else
+	mkdir $projectDir/temp_files
+	node $projectDir/index.js $workingDir
+	rm -rf $projectDir/temp_files
+fi
